Separated out the flag selector from the dalog
